+++ packages/evmos-wallet/src/wallet/wagmi/chains.ts
@@ -1,4 +1,6 @@
 import * as chains from "wagmi/chains";
+import type { Chain } from "wagmi/chains";
+import { EVMOS_CHAIN } from "../../internal/wallet/functionality/networkConfig";
 
 export const evmos = {
   ...chains.evmos,
@@ -10,6 +12,6 @@ export const evmosTestnet = {
   cosmosId: "evmos_9000-4",
 };
 
-export function getEvmosChainInfo() {
-  return evmosTestnet;
+export function getEvmosChainInfo(): Chain & { cosmosId: string } {
+  return EVMOS_CHAIN.chainId === evmos.id ? evmos : evmosTestnet;
 }