+++ packages/stateful-components/src/index.tsx
@@ -1,0 +1,2 @@
+export { StatefulFooter } from "./StatefulFooter";
+export { StatefulHeader } from "./StatefulHeader";